Changes to allow chunksizes other than 4K.
Kernel changes: The previous kernel hardcoded the 4K chunks in the ddsnap_create function.
Now, the kernel will get the chunksize bits from the ddsnapd server from the reply to the
IDENTIFY command. 

ddsnapd changes: Added a SECTOR_PER_BITS definition and the reply to the IDENTIFY message
returns the number of bits for the chunksize.