Now -o, --overwrite instead of -r, --raw and -d, --difference instead of -x, --xdelta.  Miscellaneous rewrites for clarity.  Still much to be done to bring this man page up to release quality.
Note: ddsnap and zumastor will need to be changed to match.  Since
we are the only user at the moment, I doubt think there is a need to
support the old option spelling, however we need to be sure to
commit the ddsnap and zumastor changes at the same time.