Change client->snap(num) to client->snaptag
The tag is stable across deletes, snapnum is not (because some of structs in the
the snapshot table shift position on delete).

* Mostly trivial cleanup to introduce a new function, client_snaptag, that looks
up a snapshot by the snaptag recorded in the client connection.  Failing to find
the snaptag is a fatal assertion failure that can.  We used to just close the
connection and continue in this case... bad idea.  If it happens we want to know
right now, and fix it right now.

* cleans up some nonorthogonal code that handled the snapshot status query.

* makes the various flavors of variable name used for a pointer to struct snapshot
consistent: it's always called "snapshot" now.