The first single-node non-interactive regression test works for me.
- snapshot-test.sh: the test, creates testvol on sysvg and snapshots
it repeatedly, making sure files change and appear between snapshots
as necessary.  Should run on real hardware.  This is the automated
version of the test documented in install.html  In order to succeed, a
sleep 1 was required after each zumastor snapshot command.  So the
first zumastor bug has likely been found, before the test was
submitted.  For the purposes of getting a test harness together, this
test was modified to succeed despite the slop in mounting.

- zuma-dapper-i386.sh: creates a second template based on the general
dapper install, but with the current zumastor packages installed.
Installs from the debs that resulted from the most recent
buildpackages.sh.  Preliminary (disabled) support for leaving this in
a running vm state, so the next startup doesn't need to boot.

- test-dapper-i386.sh: add executing each parameter (scripts) inside the test

- swap2sysvg.sh: set up sysvg in the existing, large swap space of the
dapper install

- test-zuma-dapper-i386.sh: like test-dapper-i386.sh, but using the
zumastored-image as a base.  Reports whether each executable script
parameter succeeds or fails.

- dapper-late.sh: add lvm2 to the default install, so sysvg can be alllocated.

- dapper-i386.sh: second bootup to run the swap2sysvg script.

- wait.sh: a simple parameter script for use as an argument in the
test-*.sh scripts.  Causes the usually non-interactive tests to halt
and wait for a return for debugging.