This patch makes the snapshot scheduling more flexible by supporting arbitrary rotation naming to create schedules other than hourly, daily, weekly, and monthly.  There are no cron jobs for these, so administrators can configure these to whatever schedule they want.
Since the rotation names are stored in the filesystem database, they follow
the same naming restriction as volumes in addition to the four standard ones.
To avoid a collision writing the rotation names to the master trigger, the
"target" and "manual" exceptions have been renamed to use slashes, which are
not allowed.