Resumable replication support:
* ddsnap transmit --resume <address> option
<address> is the byte position in the volume to resume transmitting
at.  The delta header sent downstream is modified to specify the
number of the chunks remaining, so it knows how many to expect.

* downstream zumastor receive modification
"zumastor receive start" (invoked remotely) now returns the hold
snapshot as well as the interrupted snapshot and address information
if a partial transmit has occurred.  "zumastor replication done" no
longer needs to take a snapshot number as an argument because the sent
snapshot number was recorded in the "apply" progress file.

* zumastor target daemon modification
The target daemon uses the new information returned from downstream on
"receive start" to attempt a resume if possible.  If a replication
cycle fails to complete successfully, it will continue to resume the
interrupted replication cycle unless it gets squashed upstream.  We
can add more policy later to limit the number of retries or notify the
administrator after a given number of failures with an option like
--retries option to "define target".

* dual downstream snapshots
The zumastor master now skips a snapshot number every time it creates
a new one downstream.  This leaves an unused snapnum downstream to use
to mount the volume so we can replay the journal and update the
superblock without corrupting our only copy.  ddsnap delta_apply has
been changed to reference the "from snap" rather than the origin.