Supply the ddsnapd support for snapshot revert, namely a variation on changelist generation that generates a change list against the origin, or in other words, simply a list of logical addresses of chunks that belong to a given snapshot.
This adds a new message code to distinquish between request for a
snapshot difference list vs snapshot chunk list.  Otherwise, the
same message transport is used, including the same message struct,
and one of the two snapshot fields is ignored.  This illustrates
the general idead of adding new message codes in preference to new
message flags, which helps keep the high level program flow clean,
and the message structure definitions as well.

In theory, this should not break the existing changelist generation,
autobuild/test should tell us the truth shortly.